# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CompanionGender = typeof CompanionGender[keyof typeof CompanionGender];


export const CompanionGender = {
  Male: 'Male',
  Female: 'Female',
  Other: 'Other',
  '': '',
} as const;

export interface Companion {
  name: string;
  age: string;
  gender: CompanionGender;
  relation: string;
}

export type RegistrationInputGender = typeof RegistrationInputGender[keyof typeof RegistrationInputGender];


export const RegistrationInputGender = {
  Male: 'Male',
  Female: 'Female',
} as const;

export interface RegistrationInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  fatherName: string;
  /** @minLength 1 */
  motherName: string;
  /** @minLength 1 */
  age: string;
  gender: RegistrationInputGender;
  /** @minLength 1 */
  mobile: string;
  /** @minLength 1 */
  whatsapp: string;
  /** @minLength 1 */
  village: string;
  /** @minLength 1 */
  block: string;
  /** @minLength 1 */
  district: string;
  allergy: string;
  /** @maxItems 10 */
  companions: Companion[];
}

export type Registration = RegistrationInput & {
  id: number;
  createdAt: string;
  updatedAt: string;
};

export interface Error {
  error: string;
}