# lib/api-zod/src/generated/types/companion.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CompanionGender } from './companionGender';

export interface Companion {
  name: string;
  age: string;
  gender: CompanionGender;
  relation: string;
}