# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Companion } from './companion';
import type { RegistrationInputGender } from './registrationInputGender';

export interface RegistrationInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  fatherName: string;
  /** @minLength 1 */
  motherName: string;
  /** @minLength 1 */
  age: string;
  gender: RegistrationInputGender;
  /** @minLength 1 */
  mobile: string;
  /** @minLength 1 */
  whatsapp: string;
  /** @minLength 1 */
  village: string;
  /** @minLength 1 */
  block: string;
  /** @minLength 1 */
  district: string;
  allergy: string;
  /** @maxItems 10 */
  companions: Companion[];
}